ID updates 2

+++ src/ConcreteEngine.Renderer/Core/MaterialId.cs
@@ -11,8 +11,7 @@ public readonly record struct MaterialId : IComparable<MaterialId>
     public MaterialId(int value) => Id = (ushort)value;
 
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    public ushort Index() => (ushort)(Id - 1);
-
+    public int Index() => Id - 1;
     public int CompareTo(MaterialId other) => Id.CompareTo(other.Id);
 
     public static MaterialId Empty = new(0);
